refactor: 리팩토링 및 코드 정리 (#41)
- barrel file 정리
- barrel file 추가 후 경로 수정
- 컴포넌트 경로 수정
- import 경로 수정
