+++ app/[locale]/page.tsx
@@ -4,7 +4,7 @@ import { useTranslations } from 'next-intl'
 
 import { Banner, Category, ExploreSection } from '@/components/main'
 import { useTour } from '@/hooks/tour/useTour'
-import { useLanguageStore } from '@/store/language/languageStore'
+import { useLanguageStore } from '@/store/language'
 import type { AreaBasedList } from '@/types/tour/areaBasedList'
 
 export default function Home() {
