+++ src/components/common/Badge.tsx
@@ -1,7 +1,7 @@
 import { LucideIcon } from 'lucide-react'
 
 import { badgeVariants } from '@/foundation/badge'
-import { cn } from '@/lib/cn'
+import { cn } from '@/lib'
 import { BadgeType } from '@/types'
 
 type BadgeProps = {
