+++ src/components/main/ExploreSection.tsx
@@ -7,9 +7,8 @@ import { useMemo, useState } from 'react'
 
 import { TrendingCard } from '@/components/card'
 import { FilterBadge } from '@/components/common'
-import { CONTENT_TYPE_LABEL } from '@/constants/main/contentTypeMapping'
-import { FILTER_OPTIONS } from '@/constants/main/filterOptions'
-import { useLanguageStore } from '@/store/language/languageStore'
+import { CONTENT_TYPE_LABEL, FILTER_OPTIONS } from '@/constants/main'
+import { useLanguageStore } from '@/store/language'
 import type { AreaBasedList } from '@/types/tour/areaBasedList'
 import { parseTourApiResponse } from '@/utils/tourApiParser'
 import { filterTourItems } from '@/utils/tourFilter'
