+++ src/components/map/index.ts
@@ -1,0 +1,1 @@
+export { default as OverlayBubble } from './OverlayBubble'
